batch: Extract ownership remapping
Batch ownership now contains claim metadata and selection-to-ownership
translation.

Source remapping uses lineage and line matching to move existing ownership
into a refreshed source space, which is a separate responsibility from
storing the ownership model.

This commit moves ownership remapping into batch.ownership_remapping and
updates production callers to import that owner.

Subsequent commits will move tests and architecture docs to the remapping
boundary.
